[TS migration] Migrate 'PinButton.js' component to TypeScript
